% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_db.R
\name{fetch}
\alias{fetch}
\alias{fetch_weather_synoptic}
\alias{fetch_weather_daily}
\title{Fetch subsets of weather data from the databases}
\usage{
fetch_weather_synoptic(station_ids, date_range, tzone_out = Sys.timezone())

fetch_weather_daily(station_ids, date_range)
}
\arguments{
\item{station_ids}{vector of strings; Synoptic IDs of weather stations}

\item{date_range}{vector of Dates; min and max dates to fetch weather data}

\item{tzone_out}{the timezone in which all of the dates in the database will
be read in. Dates in duckdb appear to all be in UTC. Setting tzone_out to
tz seems to be equivalent to using lubridate::with_tz(time, tz).}
}
\value{
a nested tibble with station metadata on the top level and weather
data inside of "obs"
}
\description{
Fetch subsets of weather data from the databases
}
\details{
Only a single, unified timezone is supported for now, due to how the duckdb
was written.
}
