% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db.R
\name{update_db}
\alias{update_db}
\alias{update_ghcnd_db}
\alias{update_synoptic_db}
\alias{update_unified_db}
\title{Update DBs from local files}
\usage{
update_ghcnd_db(station_equivalence_df = NULL)

update_synoptic_db(station_equivalence_df)

update_unified_db(station_equivalence_df)
}
\arguments{
\item{station_equivalence_df}{a data.frame() with IDs for stations; for now,
GHCN_ID and (Synoptic) Station_ID are assumed}
}
\value{
update_ghcnd_db() and update_synoptic_db(): This function is
typically called for its side-effect; namely, writing updates to the DB. When
station_equivalence_df is NULL, nothing is returned. If a
station_equivalence_tbl is given, then hee subset of the ghcnd_tbl mentioned
in this equivalence table is returned.

update_unified_df(): Nothing; this function is called for its side-effect, which is
writing to the DB.
}
\description{
Update DBs from local files
}
