% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weatherAndClimateDatabase-package.R
\docType{package}
\name{weatherAndClimateDatabase-package}
\alias{weatherAndClimateDatabase}
\alias{weatherAndClimateDatabase-package}
\title{weatherAndClimateDatabase: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
